web/backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.18.0
pydantic>=2.5.0
python-dotenv>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0
